+++ cpp/math/Plane.cpp
@@ -5,10 +5,10 @@ Plane::Plane() : constant(0.0) {
 }
 
 Plane* Plane::setComponents(
-	double x,
-	double y,
-	double z,
-	double w
+	float x,
+	float y,
+	float z,
+	float w
 ) {
 	this->normal.set(x, y, z);
 	this->constant = w;
@@ -16,13 +16,13 @@ Plane* Plane::setComponents(
 }
 
 Plane *Plane::normalize() {
-	double inverseNormalLength = 1.0 / this->normal.length();
+	float inverseNormalLength = 1.0 / this->normal.length();
 	this->normal.multiplyScalar( inverseNormalLength );
 	this->constant *= inverseNormalLength;
 	return this;
 }
 
-double Plane::distanceToPoint(
+float Plane::distanceToPoint(
 	Vector3 *point
 ) {
 	return this->normal.dot(point) + this->constant;
